requirements.txt
customtkinter
yt-dlp
pyperclip
Pillow